+++ src/xmldiff/input.py
@@ -27,7 +27,7 @@ def tree_from_stream(stream,
                      norm_sp=1, ext_ges=0, ext_pes=0, include_comment=1,
                      html=0):
     """
-    create internal tree from xml stream (open file or IOString)
+    create internal tree from xml stream (open file or StringIO)
     if norm_sp = 1, normalize space and new line
     """
     handler = SaxHandler(norm_sp, include_comment)
@@ -35,19 +35,18 @@ def tree_from_stream(stream,
         parser = make_parser(["xml.sax.drivers2.drv_sgmlop_html"])
     else:
         parser = make_parser()
-        # do not perform Namespace processing
-        parser.setFeature(feature_namespaces, 0)
+        # do perform Namespace processing
+        parser.setFeature(feature_namespaces, 1)
     # do not include any external entities
     try:
         parser.setFeature(feature_external_ges, ext_ges)
-        # xml.sax._exceptions.
     except SAXNotRecognizedException:
         print('Unable to set feature external ges')
     try:
         parser.setFeature(feature_external_pes, ext_pes)
-        # xml.sax._exceptions.
     except SAXNotRecognizedException:
         print('Unable to set feature external pes')
+
     # add lexical handler for comments,  entities, dtd and cdata
     parser.setProperty(property_lexical_handler, handler)
     parser.setContentHandler(handler)