+++ src/xmldiff/parser.py
@@ -47,8 +47,42 @@ class SaxHandler(ContentHandler):
         self._xpath = ''
         self._h = {}
         self._n_elmt = 0
+        self._ns_mapping = {None: [None]}
+        self._new_mappings = {}
+        self._default_ns = None
+
+    def startPrefixMapping(self, prefix, uri):
+        self._new_mappings[prefix] = uri
+        try:
+            self._ns_mapping[prefix].append(uri)
+        except KeyError:
+            self._ns_mapping[prefix] = [uri]
+        if prefix is None:
+            self._default_ns = uri
+
+    def endPrefixMapping(self, prefix):
+        ns_uri_list = self._ns_mapping[prefix]
+        ns_uri_list.pop()
+        if prefix is None:
+            self._default_ns = ns_uri_list[-1]
+
+    def _buildTag(self, ns_name_tuple):
+        ns_uri, local_name = ns_name_tuple
+        if ns_uri:
+            el_tag = "{%s}%s" % ns_name_tuple
+        elif self._default_ns:
+            el_tag = "{%s}%s" % (self._default_ns, local_name)
+        else:
+            el_tag = local_name
+        return el_tag
 
     ## method of the ContentHandler interface #################################
+    def startElementNS(self, name, qname, attributes):
+        if attributes:
+            attributes = dict(
+                [(self._buildTag(k), v) for k, v in attributes.items()])
+        self.startElement(self._buildTag(name), attributes)
+
     def startElement(self, name, attrs):
         # process xpath
         self._xpath = "%s%s%s" % (self._xpath, '/',  name)
@@ -60,10 +94,8 @@ class SaxHandler(ContentHandler):
         self._xpath = "%s%s%s%s" % (
             self._xpath, '[', self._h[self._xpath], ']')
         # nodes construction for element's attributes
-        keys = attrs.keys()
         # sort attributes to avoid further moves
-        keys.sort()
-        for key in keys:
+        for key in sorted(attrs.keys()):
             self._n_elmt += 2
             attr_node = [NT_ATTN, '@%sName' % key, key, [], None, 1, 0]
             link_node(node, attr_node)
@@ -74,6 +106,9 @@ class SaxHandler(ContentHandler):
         link_node(self._p_stack[-1], node)
         # set current element on the top of the father stack
         self._p_stack.append(node)
+
+    def endElementNS(self, ns_name, qname):
+        self.endElement(self._buildTag(ns_name))
 
     def endElement(self, name):
         # process xpath
@@ -119,7 +154,7 @@ class SaxHandler(ContentHandler):
                     0, self._h[xpath]]
             link_node(self._p_stack[-1], node)
 
-    # methods from xml.sax.saxlib.LexicalHandler (avoid dependencie to pyxml)
+    # methods from xml.sax.saxlib.LexicalHandler (avoid dependency on pyxml)
     def startDTD(self, name, public_id, system_id):
         """Report the start of the DTD declarations, if the document
         has an associated DTD.